Renamed klee/CommandLine.h to klee/SolverCmdLine.h, since this file is meant to have only solver options.
