+++ include/klee/Solver.h
@@ -10,7 +10,7 @@
 #ifndef KLEE_SOLVER_H
 #define KLEE_SOLVER_H
 
-#include "klee/CommandLine.h" // FIXME: This is just for CoreSolverType
+#include "klee/SolverCmdLine.h" // FIXME: This is just for CoreSolverType
 #include "klee/Expr.h"
 
 #include <vector>
