+++ lib/Basic/ConstructSolverChain.cpp
@@ -11,7 +11,7 @@
  * This file groups declarations that are common to both KLEE and Kleaver.
  */
 #include "klee/Common.h"
-#include "klee/CommandLine.h"
+#include "klee/SolverCmdLine.h"
 #include "klee/Internal/Support/ErrorHandling.h"
 #include "llvm/Support/raw_ostream.h"
