+++ lib/Solver/CoreSolver.cpp
@@ -10,7 +10,7 @@
 #include "STPSolver.h"
 #include "Z3Solver.h"
 #include "MetaSMTSolver.h"
-#include "klee/CommandLine.h"
+#include "klee/SolverCmdLine.h"
 #include "klee/Internal/Support/ErrorHandling.h"
 #include "klee/Solver.h"
 #include "llvm/Support/ErrorHandling.h"
