+++ tools/kleaver/main.cpp
@@ -17,7 +17,7 @@
 #include "klee/Solver.h"
 #include "klee/SolverImpl.h"
 #include "klee/Statistics.h"
-#include "klee/CommandLine.h"
+#include "klee/SolverCmdLine.h"
 #include "klee/Common.h"
 #include "klee/util/ExprPPrinter.h"
 #include "klee/util/ExprVisitor.h"
